requirements.txt
crewai
streamlit
pandas
matplotlib
python-dotenv
TextBlob
praw